rename all flywheel stuff to flywheels (squash)
