+++ src/main/java/frc/robot/RobotContainer.java
@@ -42,10 +42,10 @@ import frc.robot.subsystems.shooter.BeamBreakIO;
 import frc.robot.subsystems.shooter.BeamBreakIORev;
 import frc.robot.subsystems.shooter.BeamBreakIOSim;
 import frc.robot.subsystems.shooter.Shooter;
-import frc.robot.subsystems.shooter.flywheel.Flywheel;
-import frc.robot.subsystems.shooter.flywheel.FlywheelIO;
-import frc.robot.subsystems.shooter.flywheel.FlywheelIOSim;
-import frc.robot.subsystems.shooter.flywheel.FlywheelIOSparkMax;
+import frc.robot.subsystems.shooter.flywheels.Flywheels;
+import frc.robot.subsystems.shooter.flywheels.FlywheelsIO;
+import frc.robot.subsystems.shooter.flywheels.FlywheelsIOSim;
+import frc.robot.subsystems.shooter.flywheels.FlywheelsIOSparkMax;
 import frc.robot.subsystems.shooter.pivot.Pivot;
 import frc.robot.subsystems.shooter.pivot.PivotIO;
 import frc.robot.subsystems.shooter.pivot.PivotIOSim;
@@ -94,7 +94,7 @@ public class RobotContainer {
         aprilTagVision = new AprilTagVision(new AprilTagVisionIOLimelight("limelight"));
         shooter =
             new Shooter(
-                new Flywheel(new FlywheelIOSparkMax()),
+                new Flywheels(new FlywheelsIOSparkMax()),
                 new Pivot(new PivotIOSparkMax(), aprilTagVision),
                 new BeamBreakIORev());
         break;
@@ -112,7 +112,7 @@ public class RobotContainer {
         aprilTagVision = new AprilTagVision(new AprilTagVisionIO() {});
         shooter =
             new Shooter(
-                new Flywheel(new FlywheelIOSim()),
+                new Flywheels(new FlywheelsIOSim()),
                 new Pivot(new PivotIOSim(), aprilTagVision),
                 new BeamBreakIOSim());
         break;
@@ -130,7 +130,7 @@ public class RobotContainer {
         aprilTagVision = new AprilTagVision(new AprilTagVisionIO() {});
         shooter =
             new Shooter(
-                new Flywheel(new FlywheelIO() {}),
+                new Flywheels(new FlywheelsIO() {}),
                 new Pivot(new PivotIO() {}, aprilTagVision),
                 new BeamBreakIO() {});
         break;
