+++ src/main/java/frc/robot/subsystems/shooter/Shooter.java
@@ -3,7 +3,7 @@ package frc.robot.subsystems.shooter;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.subsystems.leds.Leds;
-import frc.robot.subsystems.shooter.flywheel.Flywheel;
+import frc.robot.subsystems.shooter.flywheels.Flywheels;
 import frc.robot.subsystems.shooter.pivot.Pivot;
 import frc.robot.util.VirtualSubsystem;
 import frc.robot.util.loggedShuffleboardClasses.LoggedShuffleboardBoolean;
@@ -12,7 +12,7 @@ import org.littletonrobotics.junction.Logger;
 public class Shooter extends VirtualSubsystem {
   private final double kDistSensorRangeWhenNoteInches = 2.5;
 
-  private final Flywheel flywheel;
+  private final Flywheels flywheel;
   private final Pivot pivot;
   private final BeamBreakIO beamBreakIO;
 
@@ -20,7 +20,7 @@ public class Shooter extends VirtualSubsystem {
   private final LoggedShuffleboardBoolean beamBreakWorkingEntry =
       new LoggedShuffleboardBoolean("BeamBreak Working", "Shooter", true);
 
-  public Shooter(Flywheel flywheel, Pivot pivot, BeamBreakIO beamBreakIO) {
+  public Shooter(Flywheels flywheel, Pivot pivot, BeamBreakIO beamBreakIO) {
     this.flywheel = flywheel;
     this.pivot = pivot;
     this.beamBreakIO = beamBreakIO;
